requirements.txt
crewai
crewai_tools
streamlit
python-dotenv
groq
youtube-transcript-api
